Extracted AppSettings and ConnectionStrings classes into their own files and removed ConfigurationHelper file.
