+++ Application/EdFi.Ods.AdminApp.Management/Helpers/AppSettings.cs
@@ -29,26 +29,4 @@ namespace EdFi.Ods.AdminApp.Management.Helpers
         public string OptionalEntropy { get; set; }
         public string Log4NetConfigPath { get; set; }
     }
-
-    public class ConnectionStrings
-    {
-        public string Admin { get; set; }
-        public string Security { get; set; }
-        public string ProductionOds { get; set; }
-
-        public string GetConnectionStringByName(string databaseName)
-        {
-            switch (databaseName)
-            {
-                case CloudOdsDatabaseNames.Admin:
-                    return Admin;
-                case CloudOdsDatabaseNames.Security:
-                    return Security;
-                case CloudOdsDatabaseNames.ProductionOds:
-                    return ProductionOds;
-                default:
-                    return null;
-            }
-        }
-    }
 }
